Add Vagrant sandbox
Vagrant box provides an Ubuntu (14.04) VM with YDK-Py pre-installed. The
`samples` and `projects` directories are shared between the host and
the Vagrant box.  Any changes to those directories are seen in both
environments.  Any other data in the Vagrant box is isolated from the
host and will be lost if the Vagrant box is destroyed.